bugfix
for every blocked tile, check if the piece blocking it is the piece being moved. (changed from bishop only). now also checks if king is moving from white to white or black to black.
enemy color is no longer a safe square from check if the enemy king is able to defect and move in the same turn.